Added weapons and firing arcs; changed facing to double for higher precision

+++ src/starships/entities/IMovable.java
@@ -3,8 +3,8 @@ package starships.entities;
 public interface IMovable {
     enum turningDirections {LEFT, RIGHT}
 
-    int getFacing();
-    void setFacing(int newFacing);
+    double getFacing();
+    void setFacing(double newFacing);
     void moveForward();
      void turn(turningDirections direction);
 }
